routing functionality moved from app to file
